backend/quiz_game_table/fetch questions/requirements.txt
google-cloud-firestore
Flask
flask-cors
boto3
pyyaml